deleted some bugs and display error, and changed some warning msg

+++ ProcurementSystem/ProcurementSystem/PurchaseRequest.cs
@@ -125,22 +125,25 @@ namespace ProcurementSystem
             errorEditMsg.Text = "";
             errorMsg.Text = "";
             errorSearchMsg.Text = "";
-
             for (int i = 0; i < purchaseRequestList.Items.Count; i++)
             {
                 try
                 {
-                    if (purchaseRequestList.GetItemChecked(i))
+                    MySqlCommand checkNew = new MySqlCommand("Select Status where RequestNo ='"+purchaseRequestList.Items[i]+"';",cnn);
+                    string checkStatus = checkNew.ExecuteScalar().ToString();
+                    if (purchaseRequestList.GetItemChecked(i) && checkStatus == "NEW")
                     {
                         del = new MySqlCommand("Delete from PurchaseRequest where RequestNo = '" + purchaseRequestList.Items[i].ToString() + "'", cnn);
                         del.ExecuteNonQuery();
+                    }
+                    else {
+                        errorMsg.Text = "Some of the request is handled, Cannot Edit or delete!";
                     }
                 }
                 catch (Exception ex)
                 {
                     Console.WriteLine(ex.ToString());
                     errorMsg.Text = "Some of the request is handled, Cannot Edit or delete!";
-                    break;
                 }
                 //purchaseRequestList.Items.RemoveAt(i);
                 finally
